requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
langchain>=0.1.0
langchain-core>=0.1.16
langchain-openai>=0.0.5
openai>=1.6.0
python-dotenv>=1.0.0
pyairtable>=2.0.0
requests>=2.31.0
python-multipart>=0.0.6
